refactor: consolidate duplicate dashboard types to shared location
Extract WeeklyTransfer, TopPositionPlayer, TopPositionPlayers, LeagueAverages,
PositionMover, and PositionMovers types to shared Dashboard.ts file to eliminate
duplication across components and composables.
